+++ plugins/fcitx5/binding/gen_bind.sh
@@ -1,0 +1,4 @@
+#!/bin/bash  
+set -e  
+
+cbindgen --config cbindgen.toml --crate lipilekhika-fcitx5-binding --output include/lipilekhika_typing.h