+++ packages/python/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Unreleased (python-lib@v1.0.1)
 
 - Added `get_script_krama_data`
+- Added `context_length` to `take_key_input` return value